backend/requirements.txt
amqp==5.3.1
async-timeout==5.0.1
bcrypt==4.0.1
billiard==4.2.2
blinker==1.9.0
celery==5.3.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
greenlet==3.2.4
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.0
MarkupSafe==3.0.3
packaging==25.0
pdf2image==1.16.3
Pillow==10.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.7
PyJWT==2.10.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
redis==5.0.1
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
